+++ src/main/java/com/kaiki/minechatcorrect/mixin/ScreenInvoker.java
@@ -2,7 +2,7 @@ package com.kaiki.minechatcorrect.mixin;
 
 import net.minecraft.client.gui.components.Renderable;
 import net.minecraft.client.gui.components.events.GuiEventListener;
-import net.minecraft.client.gui.components.TabOrderedElement;
+import net.minecraft.client.gui.narration.NarratableEntry;
 import net.minecraft.client.gui.screens.Screen;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Invoker;
@@ -10,5 +10,5 @@ import org.spongepowered.asm.mixin.gen.Invoker;
 @Mixin(Screen.class)
 public interface ScreenInvoker {
     @Invoker("addRenderableWidget")
-    <T extends GuiEventListener & Renderable & TabOrderedElement> T mineChatCorrect$addRenderableWidget(T widget);
+    <T extends GuiEventListener & Renderable & NarratableEntry> T mineChatCorrect$addRenderableWidget(T widget);
 }